new keywords/operator for domain (contains, not contains) to use in attributes
logic similar to "in" and "not in" operators but with a twist :)
first element is field name of field list container - expected value in this field is comma separated list of fields that will be affected (without spaces)
second is keyword/operator - "contains" and "not contains" which is self-explanatory
last element is field name of the field that you want to affect

eg. <field name="affected_field" attrs="{'invisible':['field_list_container', 'contains', 'affected_field']}"/>
where field_list_container is field of type Char that can also be computed and we can make three separate fields for invisible, required, readonly and the values in container can be  eg. "field_1,field_2,feld3"